irqchip/rtl8380: Add driver for Realtek rtl8380 family of SoCs
This patch adds a new driver for the interrupt cascade found on MIPS based
switch SoC produced by Realtek.

The intc is wired to 4 of the MIPS cores interrupt lines. The intc has a
maximum capacity of 32 internal lines, only 12 are actually wired up.

There are 4 IRRs (Interrupt Routing Register) that define which of the
soc IRQ lines is wired to what MIPS IRQ lines. These values appear to be
static and are taken directly from the SDK.

The actual IRQ silicon block is very trivial and has no enable/disable
registers. Apart from the IRR there are only two further registers.

* GIMR - IRQ masking
* GISR - IRQ status

Signed-off-by: Birger Koblitz <mail@birger-koblitz.de>
Signed-off-by: Bert Vermeulen <bert@biot.com>
Signed-off-by: John Crispin <john@phrozen.org>