MIPS/realtek: add Realtek switch SoC support
The RTL8380/8390/9300 family of SoCs are Realtek switches with an embedded
MIPS core.

* RTL8380 - 500MHz 4kce single core - 1Gbit ports and L2 features
* RTL8390 - 800MHz 4kce single core - 1Gbit ports and L2 features
* RTL9300 - 800MHz interaptiv multi core - 1/2,5/5/10 Gbit ports and L3 features

These switches, depending on the exact part number will have anywhere
between 8 and 64 ports. The MIPS core is wired to a cpu port which has a
tagging feature allowing us make use of the DSA subsystem. The cores are
very primitiv in certain areas. Several silicon blocks are very basic. The
INTC only has 2 + 4 registers. There is not real controllable clock and
power hierachy tree.

The switch functionality is mmio mapped via a large MFD region.

The SoCs have the following peripherals
* ethernet
* switch
* uart - ns16550a
* spi-flash interface
* wdt
* gpio
* a bus interface to connect up to 16 external gpio controllers

The code was dervived from various v2.6 based SDKs.

This patch and the following DTS patch allow us to boot RTL8380 units with
basic support. Most of the other drivers are already written and functional
and work to get them upstream is already in progress.

Signed-off-by: Birger Koblitz <mail@birger-koblitz.de>
Signed-off-by: Bert Vermeulen <bert@biot.com>
Signed-off-by: John Crispin <john@phrozen.org>